MB-35908: Compare cas instead of seqno when deduping setVBState
The flusher currently compares the seqnos of set_vbucket_state
items to determine which is the most recent vbucket state (i.e. the
one that should be persisted). This is an issue as set_vbucket_state
operations do not update the lastBySeqno of the vBucket. This means
that subsequent set_vbucket_state items may have the same seqno.
It's not valid to udpate the lastBySeqno of the vBucket as the
replica seqnos will be different, consider setting the failover
table on DCP stream acceptance, and DCP Consumers may expect to see
a seqno that will never be sent.

Compare the CAS of the set_vbucket_state items to determine which
is the most recent instead of the seqno as the CAS will be unique.

Change-Id: Id17e03b635710c04e195e3ff77e407a065e112cd
Reviewed-on: http://review.couchbase.org/114822
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>