MB-35979: Ensure SyncWrites check expiry of committed item
Writes with cas respond differently if the existing item is logically
non-existent (e.g., expired) vs exists with different cas.

Prior to this patch, processSetInner checked the expiry of the stored
value which is about to be modified. However, in an ephemeral bucket
this stored value may be a completed prepare - the expiry of the
prepare is not necessarily related to the expiry of the committed item.

This patch ensures it is always the expiry time of the committed item
which is checked.

Change-Id: I4a3200cbc363e42f2fc9e5a8009b1a31fd91fdc8
Reviewed-on: http://review.couchbase.org/114825
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>