MB-35979: Ensure SyncReplaces check for committed values
A SyncReplace should only succeed if there is a committed value which
is not expired or deleted.

This patch fixes a bug affecting ephemeral buckets allowing
SyncReplaces even if there is no committed value in the presence of a
completed prepare. Replace previously only tested the state of the
passed stored value - this is insufficient as the passed value may be a
completed prepare, which does not provide information on whether a
_committed_ value exists.

Change-Id: I219071b95b93d0f7abcc3b73a1d101f523ad04ed
Reviewed-on: http://review.couchbase.org/114826
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>