MB-35979: Ensure SyncAdds check for committed values
A SyncAdd should only succeed if there is no committed value (or the
committed value is logically non-existent e.g., expired).

This patch fixes a bug affecting ephemeral buckets allowing SyncAdds
even if there is a committed value. Process add previously only tested
the state of the passed stored value - this is insufficient as the
passed value may be a completed prepare, which does not provide
information on whether a _committed_ value exists.

Change-Id: I2adc5f85733056f254458142ce035a645971f029
Reviewed-on: http://review.couchbase.org/114827
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>