MB-33739: Convert VBConn connHandler member to reference
Promoting the weak_ptr from the vbToConns to a shared_ptr is
very (cache) contentious as it is called by different threads so
change this to a reference now that we have introduced ConnStore
to reduce the likelihood of ConnHandler ownership bugs.

This patch makes no noticeable change in perf stats. The baseline has
moved greatly since first attempting to do this. Noted a lack of cache
contention on promoting pointers to the DcpProducer in a toy build with
this patch though.

Change-Id: Ib49a2be5b82cbab4b8b8f5e92f002938ee757c1b
Reviewed-on: http://review.couchbase.org/114829
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>