MB-36027: Use a single buffer for GET reponse [SSL]
A get response looks like:

   +-------------------------------+
   |       24 byte header          |
   +-------------------------------+
   |        4 byte flags           |
   +-------------------------------+
   |  n bytes key (if requested)   |
   +-------------------------------+
   |        n byte value           |
   +-------------------------------+

Before this change we would put each of these segments
in an IO vector and use sendmsg to transfer all of
it back to the client. This was working great for
our plain connections, but when using SSL it had
an unexpected negative sideeffect: It would generate
a separate TLS frame for each segment, even if we
could fit it all in a single frame.

To work around the problem we'll format the entire
packet in a single buffer if the payload (everything
except the packet header) is less than 4k. (It might
be that we want to reduce/increase the 4k limit
depending if the memory copying cost compared to
the extra hashing and increased data to send on
the wire)

Change-Id: I76cc2a2c4d20f8073e34ceb59b48613d696f9f09
Reviewed-on: http://review.couchbase.org/114930
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>