MB-36014: Fix tracking number of cursors in checkpoint for replica
Prior to this change, if any cursors other than the persistence cursor
were created /before/ receiving a snapshot from the active, the cursors
would be reset to the first checkpoint, and would increment the number
of cursors in that checkpoint again.

The major result of this was that affected checkpoints, and all
following checkpoints, would never be removed as they would forever be
seen as "referenced" because of the erroneous second increment.

Fixed by only incrementing if resetting the cursor to a different
checkpoint.

Change-Id: Iaad476cf2ebeca53597856022848ad10a8811276
Reviewed-on: http://review.couchbase.org/114960
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>