MB-36067: Make CheckpointManager::clear set correct checkpoint ID
A small omission lead to clear_UNLOCKED always setting the new
checkpoint ID as 1, despite an apparent attempt to set it as 0 if the
vbucket is not active.

Seemingly, the error was introduced when the method was refactored in
http://review.couchbase.org/#/c/98271/ . No consequences of this have
been seen, but based on the noted patch it seems reasonable to restore
the original logic.

Change-Id: I21197369f130b0283dddb42e30025186bb9b57f3
Reviewed-on: http://review.couchbase.org/115027
Tested-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>