Use vbstate operator<< for logging
Currently we are paying the cost to convert the vbstate to json
and then back into a string just to attempt to log it. This is done
before the log level is checked and this function is called in every
flush. On my machine, roughly 3.5% of the time spent flushing a
single vBucket was spent converting the vbstate (~1000 durable
writes per second, 2 nodes, 1024 vBuckets).

Change-Id: Iec5769b69be251aa5e062ad5e88cbb56c0d4a069
Reviewed-on: http://review.couchbase.org/115028
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>