Add should_log check to EP_LOG_FMT macro
We don't want to pay the cost of formatting any arguments by
calling functions (e.g. converting the vbstate to json then to
a string) if we are not actually going to log. Add the should_log
check to the macro (in addition to the check inside
BucketLogger::log) so that any uses of the macro will not spend
unnecessary time formatting arguments.

Change-Id: Id3ca0db1c608065f1663a6b71bb71d4d9c9c8004
Reviewed-on: http://review.couchbase.org/115030
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>