MB-32362: Delete old files in testapp_persistence test SetUp
We currently have a sporadic failure in the tests in the
ConsistentStateAfterShutdown tests in this file that could be
caused by not deleting the old files and warming up old data as
the tests rely on specific seqnos. Delete the old files for every
test run (instead of just when the test suite is created) to ensure
that we have the correct seqnos.

Change-Id: I963995b850db16668f63bdb12e32cf1f1da8f150
Reviewed-on: http://review.couchbase.org/115096
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>