MB-36102: Add resolved prepares to trackedWrites at Active->NonActive
If we are about to transition from active to non-active then we need
to ensure that the DM state is consistent with the HashTable as we use
it to create a PDM. If we were to process the queue then this node
would get out of step with the new active and need to rollback (or
potentially have two different items with the same seqno). Similarly,
we cannot simply remove the items in the queue as the node will need
the prepare in trackedWrites if it becomes a replica. As such, we
should add the prepares from the resolvedQueue back to trackedWrites
when we transition away from active.

Change-Id: I2478fa3334d2495e1431dd785555f45802735104
Reviewed-on: http://review.couchbase.org/115129
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>