MB-33463: Change initScan doc count to distance between seqno's
Magma use to pass in docCount (same as dbinfo.doc_count) to scan to
indicate the number of documents to scan. What couchstore_changes_since
does is count the number of document changes since the startSeq. Magma
does not know and can't figure out exactly the number of documents
that have changed so the best it can do is set doc_count to the distance
between the startSeqno and the highest seqno seen. The doc_count will be
impacted by the number of documents that are removed during batch
processing which are duplicates and is solely based on application
usage.

Change-Id: I6c1d6a6024776fe3ff77ceb37a2c7c97026337a7
Reviewed-on: http://review.couchbase.org/115151
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>