MB-32362: Allow flusher to transition from Initializing to Pausing
Currently the flusher can only transition to Pausing from the
Running or Paused states. If we attempt to pause the flusher before
it runs for the first time then it will be in the Initializing
state and this transition will not be valid. Allow the state of the
flusher to transition from Initializing to Pausing.

Change-Id: Id3e872dec57efc9fa4f3753e385295defca24794
Reviewed-on: http://review.couchbase.org/115182
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>