Recommit: Rename vbucket_state::high[CS|PS] to persisted[CS|PS]
The on disk highCompletedSeqno and highPreparedSeqno lag behind the
one that is stored in memory and are dealt with differently when it
comes to things such as disk snapshots. It can often be confusing
as to which completedSeqno/preparedSeqno one is referring so
renaming them to persisted[CS|PS] makes this clearer.

Change-Id: Ie8c2fc403b9c905471a81f98db6380e61f730649
Reviewed-on: http://review.couchbase.org/115323
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>