Rename vbucket_state::high[CS|PS] to persisted[CS|PS]
The on disk highCompletedSeqno and highPreparedSeqno lag behind the
one that is stored in memory and are dealt with differently when it
comes to things such as disk snapshots. It can often be confusing
as to which completedSeqno/preparedSeqno one is referring so
renaming them to persisted[CS|PS] makes this clearer.

Change-Id: Idef2923075452658f3318ca54c3e072e5d59aff2
Reviewed-on: http://review.couchbase.org/115215
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>