Add additional logging when we warmup a DurabilityMonitor
Currently we log useful information when we create a vBucket such as
the high seqno and persisted snapshot ranges. We do not log the HCS
or HPS warmed up as this is done in a different warmup phase after
we have created the VBucket. Add some extra logging as these values
are useful when determining if we have warmed up correctly.

Change-Id: Id97a3795035269579b6607fed59202ae462c1371
Reviewed-on: http://review.couchbase.org/115252
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>