MB-36112: Do not move PCS in disk snapshot
We should not move the persistedCompleteSeqno in a disk snapshot as
this may break our warmup optimization. We currently move it
whenever we receive an abort but due to de-dupe and out of order
completion we could end up receiving a prepare followed by an abort
for a different key. If we would ever be promoted, whilst we would
have lost data because we have an incompleted snapshot (we would see in
the first disk snapshot a completion for the first prepare), we want to
warmup and recommit the prepare as we will then be at a consistent point
up to the Abort. Also, this will prevent the PCS from being greater than
the persistedPreparedSeqno(PPS) on disk when we persist a partial disk
snapshot. This isn't a problem, but could confuse anybody looking at the
stats as the PCS logically should be lower than the PPS.

Change-Id: I96524ab01789e82cb2906381f38f3425e26b99fb
Reviewed-on: http://review.couchbase.org/115335
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>