MB-36094: Flush PCS at end of snapshot instead of checkpoint
Currently we are flushing the PCS value from Disk checkpoints at the
end of the checkpoint. This is an issue as if we receive a Disk
checkpoint then no other items we will not persist our PCS as we have
not completed our checkpoint. Instead, we need to persist our PCS at
the snapshot end to ensure that we persist it when we have a complete
snapshot.

Change-Id: I59fd3d261174d6f325bbe69e82100792b3ec77d5
Reviewed-on: http://review.couchbase.org/115346
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>