MB-36075: Revert "MB-35631: Do not skip checkpoint_end ops for persistence cursor"
This reverts commit 69044aeef5ef670fee9a3b74d739fd03e304990e.

This patch was originally created to ensure that we would always
see a checkpoint end when we flush a vBucket. This was required to
ensure that we will always flush a HCS value to disk. This method
of flushing the HCS value has now been changed to instead flush the
value in the flusher when it finds the last item in the snapshot. As
such, we can revert this patch which is causing a throughput
regression due to lingering cursors allowing checkpoint expelling to
run instead of dropping the checkpoint which forces the residency
ratio to drop as we spend longer evicting items whilst we free memory.

Change-Id: I9da1fb3f3bdcde88cd33158e690e637029ce7fd1
Reviewed-on: http://review.couchbase.org/115347
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>