MB-32362: Check keys don't exist from correct seqno
The test writes keys 1 to 100 with a write to the document "high"
in between with the current key number. If the test were to stop after
writing a number, say 2, then the value of the "high" document should
be 1. In this case, we are currently checking and expecting the
document 2 to NOT exist as we start checking from
highNumber (1) + 1 = 2. The comments in the test elude to this being
valid but the check is incorrect. Check instead from highNumber + 2
as highNumber + 1 may exist.

Change-Id: If3fe310164362edb6b134ceebb4cf709a05b06f3
Reviewed-on: http://review.couchbase.org/115353
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>