Update 'num_*_threads' config params to reflect configured thread count
The four config parameters which specify the number of
reader/writer/auxIO/nonIO threads are defaulted to zero, which means
"auto configure based on the number of CPUs in the system". However,
after determing the number of threads to use, the config parameters
are left at '0', which is (a) potentially misleading and (b) means
that calling setNumXXXThreads(getNumXXXThreads()) isn't idempotent -
and instead sets the thread count to zero which can result in
deadlock.

Address this by updating the actual configured thread counts during
bucket initialization (after we have calculated the actual thread
counts).

Change-Id: I2f6343a98fb215d6b1e2742993250215ab87f98d
Reviewed-on: http://review.couchbase.org/115420
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>