MB-34189: Allow durability_timeout_task_interval to be changed dynamically
The config parameter 'durability_timeout_task_interval' is defined to
be dynamic - it can be changed at runtime, however it is not handled
in setParam so cannot actually be changed.

Add to the set of handled keys in setFlushParam(), and add a listener
to DurabilityTimeoutTask to be notified if the value changes.

Change-Id: I69c22251a6f0e94dfffa5489cf254b2a9ebb9e52
Reviewed-on: http://review.couchbase.org/115421
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>