Prepare Command Contexts for OoO [1/n]
When operating in OoO mode the underlying buffer for the
request may change the first time the command returns EWOULDBLOCK
as we'll try not to copy the buffer if we don't need our own
copy. This means that we can't keep pointers into the packet as
members in the command context (but fetch them from the cookie
every time).

Change-Id: I4319d28ac00b8f6e92a01aba426999f875d4749c
Reviewed-on: http://review.couchbase.org/115562
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>