Prepare Command Contexts for OoO [2/n]
When operating in OoO mode the underlying buffer for the
request may change the first time the command returns EWOULDBLOCK
as we'll try not to copy the buffer if we don't need our own
copy. This means that we can't keep pointers into the packet as
members in the command context (but fetch them from the cookie
every time).

Change-Id: Ice6b62a17609ffc119aa10d6016c980b8a301c2f
Reviewed-on: http://review.couchbase.org/115567
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>