Prepare Command Contexts for OoO [2/n]
When operating in OoO mode the underlying buffer for the
request may change the first time the command returns EWOULDBLOCK
as we'll try not to copy the buffer if we don't need our own
copy. This means that we can't keep pointers into the packet as
members in the command context (but fetch them from the cookie
every time).

Change-Id: Id868da0a64f99d360f88e2e386c74c39cd482397
Reviewed-on: http://review.couchbase.org/116236
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>