Recommit: Make checkpoint seqnos weakly monotonic
Seqnos in Checkpoints are currently not weakly monotonic as some
meta items are enqueued with lastBySeqno + 1 whilst others are not.
If we were to see a set_vbucket_state followed by a checkpoint_end
then we would see our seqno go back by 1.

Enqueue all meta items with lastBySeqno + 1 to make checkpoint seqnos
weakly monotonic.

When this patch was previously committed a monotonic exception was
thrown by the ActiveStream::lastReadSeqno when it was attempting to
snapshot items for a given cursor. This occurred because the initial
stream request with non-zero seqno planted the cursor at a
checkpoint_end item (as this was a non-zero stream we had already
streamed the item at this seqno). We then attempted to re-stream the
item which caused the monotonicity exception to throw. This issue was
caused by bumping the seqno of the checkpoint_end item. To fix this,
when we call Checkpoint::getHighSeqno(), we should return the seqno of
the checkpoint_end - 1 for closed Checkpoints (i.e. the seqno of the
last real item).

Change-Id: I502bdde9fed4e6a53b9fb65787863680934a05a7
Reviewed-on: http://review.couchbase.org/115846
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>