SeqnoAck - don't search all StreamContainers for opaque
Currently we are searching the StreamContainers for every vBucket
until we find the one that matches our opaque when processing a
seqno ack. This is inefficient as we have already found the correct
StreamContainer for our vBucket.

Just search the StreamContainer we have already found for the given
vBucket for the stream with the correct opaque.

Change-Id: I92494691ff02eb629c0c1d8ce826de9550e0e0e1
Reviewed-on: http://review.couchbase.org/115579
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>