MB-36227: Make EPStats cacheline padded
One of the last members of EPStats (memUsedMergeThreshold) is hit
whenever we allocate or deallocate memory. This value currently shares
a cache line with the start of the Configuration object which contains
a mutex that is written to (to acquire it) every time we want to read
something from the configuration (at least every time we call
DcpProducer::step/maybeDisconnect). Remove the false sharing by making
EPStats cacheline padded.

Change-Id: I47763c72210a2eeb1ceb0a025836d86541c35c8b
Reviewed-on: http://review.couchbase.org/115688
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>