MB-36227: Don't write to pendingVBs unless necessary
Currently the DurabilityCompletionTask is writing false to
pendingVBs[vbid] for every vBucket it does not do any work on. This
will cause us to dirty a cache line and make it slower to schedule
the DurabilityCompletionTask for a vBucket that wishes to write to
the same cache line(s).

Only write false to the cache line if we can transition from state
true to false (i.e. work is being done).

Change-Id: I700a06395188868d7e07829f56d3321652264e0d
Reviewed-on: http://review.couchbase.org/115694
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>