MB-36261: Do not erase keys from indexes of open ckpts when expelling
Whilst it is okay to expel items from currently open checkpoints we
cannot expel the keys from the indexes as they are required to
correctly prevent de-dupe. We do not allow de-dupe to occur if one of
the items is related to durability. If we had previously expelled an
item from the index of an open checkpoint then we may inadvertently
send to the replica a snapshot containing a committed item and a
mutation. This would cause the replica to start throwing exceptions.

Do not erase the keys from the indexes of open checkpoints as this will
cause the replica to start throwing exceptions if we do not open a new
checkpoint correctly. Furthermore, correct the queueDirty logic for
when we attempt to de-dupe an item that has been expelled. We should
not be unconditionally allowing items into a checkpoint if the previous
one has been expelled. We should only do this if they can be deduped.

Change-Id: I03fe9b28b538985ad5f821fe9a4ba0b59b23f214
Reviewed-on: http://review.couchbase.org/115757
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>