Add config setting for max # of concurrent cmds
This number specifies the maximum number of cookies each
connection may have running at the same time. When
the connection reach the maximum number it has to wait
for commands to complete before starting a new one.

This changeset only adds the tunable and allows for
dynamic reconfiguration of the variable to minimise
the changeset which use it.

Change-Id: Ifa88bd2400580a273cb1a08cd9d25a23cbc42a1b
Reviewed-on: http://review.couchbase.org/115810
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>