Remove default parameters for McbpValidators
The fields in the packet header should be specified
explicitly to make it easier to see which datatypes
the packet allows for (without having to jump to the
header).

There is no such thing as a "default" values when
it comes to the fields the method is checking and
takes as parameters.

Change-Id: I408d610317fa1de6ad49bbbcfd6837eda14f12ea
Reviewed-on: http://review.couchbase.org/115822
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>