MB-36262: Move ADM::State to durability_monitor_impl.h
When we convert the ADM to a PDM we need to transfer the status of
our SyncWrite objects. To pass and use the ADM into the constructor
of the PDM, the PDM needs to know about the ADM::State. Move it to
durability_monitor_impl.h where the PDM::State resides for the same
(inverse) reason.

Change-Id: I8f5320e969f3fa69418c583357f4c484a99bfc8a
Reviewed-on: http://review.couchbase.org/115860
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>