MB-36262: Split SyncWrite into active and passive implementations
Currently the SyncWrite object is used in both the Active and Passive
DurabilityMonitor's. The majority of the state in the SyncWrite
object is there solely for the use of the Active DurabilityMonitor.
We need to transition the correct state from SyncWrites in an Active
DurabilityMonitor to the Passive DurabilityMonitor. To allow us to do
this without having to worry about potential Active only states in the
Passive DurabilityMonitor split the SyncWrite class into two so that we
can "drop" the Active only state on transition to a Passive
DurabilityMonitor.

Change-Id: I24378376b7bf844de6a69c2d8592c80c83e88dfa
Reviewed-on: http://review.couchbase.org/115924
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>