MB-36262: Remove complete writes from PDM instead of up to HCS
Currently we are removing SyncWrite objects from the PDM when they
have a seqno that is lower than both the HCS and HPS. This is
incorrect as they may not yet have been completed.

Instead, as the HCS moves inconsistently within a disk snapshot, we
should just remove completed prepares from the start of trackedWrites
up to the HPS. As we can only move the HPS after receiving a complete
snapshot, all of the completed writes in trackedWrites should be at
the beginning.

Change-Id: I563bd5784bd6db0a95b2023a40712bb4db7f680a
Reviewed-on: http://review.couchbase.org/115838
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>