MB-36227: Use VBReadyQueue in DurabilityCompletionTask
The DurabilityCompletionTask currently uses a bespoke scheduling
method with a vector of atomic bools to determine which vBuckets
it has been scheduled to run on. The idea motivating this was to
reduce contention on any given lock. One drawback that this task
scheduling method has though is that we have to do some work for
every vBucket to determine if we need to perform our given task.
In this case we are performing an atomic read and a steady_clock
now() call. This adds up which means that the time it takes for
us to run the DurabilityCompletionTask for a single vBucket is
dominated by checking if we need to run for other vBuckets and
seeing if we should yield to other tasks.

Given that the work profile of this task is very similar to that of
the ActiveStreamCheckpointProcessorTask which uses the VBReadyQueue,
use the VBReadyQueue for the DurabilityCompletionTask too. If we see
that the scheduling/running of the task is not as performant as it
should be in future profiling then it is worth changing for both of
them.

Change-Id: If2bc46e93f40abca657c55b55c5d7102a7faea4e
Reviewed-on: http://review.couchbase.org/115938
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>