MB-23086: Add setMaxDataSize method for changing the quota
Add a method that aims to consolidate various bits of existing
code that handle setting the quota. The method's aim is to
set the quota and then update other parts of the system with
respect to the new value, such as setting the water mark values
and updating expel triggers.

This patch is a re-factoring exercise, but future changes for
MB-23086 will add to this method.

Note that the changes to ep_testsuite_basic were initially made
without any other code changes to further validate the new changes
against the current quota logic.

Change-Id: Idf9dfafe22ab85f166a6d70d0e0de52f629a79d2
Reviewed-on: http://review.couchbase.org/115833
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>