Ensure all stats in FileStats & KVStoreStats reset correctly
Modify FileStats and KVStore stats to ensure that all elements
(counters and histograms) are reset correctly when the reset() method
is called.

Remove the manually-specified default ctors - all the members of the
class initialise themselves in their own ctor already.

Also de-inline the reset() methods given they are not small and are
not performance-sensitive.

Change-Id: I3d4588651dfb64f2be1302b2de1ffaed5ab432f2
Reviewed-on: http://review.couchbase.org/116009
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>