MB-36301: Partial Revert "MB-35889: Purge Checkpoint key indexes during expel and state change"
1) Partially reverts commit 581e5750f (all regressing functionality is
   reverted)
2) Fully reverts commit 22b6ac36 to avoid a test instability reoccurring

This commit is where the performance regression of MB-36301 occurs.
When setState clears the key and meta indexes it is done with the
checkpoint manager locked, increasing the wait time for other
connections writing to the VB.

This is a partial revert because left behind are a few useful parts
of the original commit.

* the more detailed checkpoint stats
* sizing CheckpointIndexKeyNamespace as uint8_t

Note that this commit includes full revert of 22b6ac36 as a revert by
itself will re-introduce a test instability.

Change-Id: I196e755251eeecc7b07488bccbbe71b62c724bee
Reviewed-on: http://review.couchbase.org/116073
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>