Decrement old checkpoints cursor count when resetting Cursors
The code previously avoided decrementing the Cursor counts for old
checkpoints when restting Cursors as we could have deleted them
already when we call clear_UNLOCKED. Modify clear_UNLOCKED to swap
the checkpointList with a new one to keep around the old checkpoints
until we have had time to clear them. This makes it easier to debug
Checkpoint Cursor counting as all increments should now be paired
with a decrement.

Change-Id: I9676e5fe8cba8696450f4cc214b4d3694b25936d
Reviewed-on: http://review.couchbase.org/116079
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>