MB-36361: RAII CheckpointCursors in regards to ref counting
Original bug:

When we do a rollback to zero we reset the vBucket by deleting the old
one and creating a new one. As part of doing this, we copy the cursors
from the old CheckpointManager to the new one. The new
CheckpointManager should only have a persistence cursor at this point.
When we copy the persistence cursor from the old CheckpointManager we
don't decrement the count for the old persistence cursor. This causes
us to double reference the first Checkpoint which will cause expelling
to throw an exception.

Solution: RAII CheckpointCursors in regards to ref counting

When we create and destroy CheckpointCursor objects we should
increment and decrement the ref count for the pointed to Checkpoint
automatically. As removeCursor doesn't delete a cursor (it just
invalidates it, currently by setting the currentCheckpoint to the end
of the list) we need to only do this if the CheckpointCursor is valid.

Change-Id: I65ddcc2e3004435bff68b106d5aba72c776c3a1f
Reviewed-on: http://review.couchbase.org/116080
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>