MB-36338: Ensure keyIndex entries are invalidated correctly
When expelling items from a checkpoint the corresponding keyIndex
entries are invalidated. This process marks whether the entry was for a
SyncWrite op (Prepare, Abort, Commit) or a normal operation (Mutation,
Deletion, etc).

To determine this, the keyIndex entry's iterator is dereferenced to
find the item. However, during expelling the value of the last item to
be expelled is swapped with that of the dummy item. This is done to
ensure the checkpoint will start with a dummy item after expelling,
/without/ invalidating iterators pointing to the last item to be
expelled.

As the dummy value is not a SyncWrite, the keyIndex entry for the last
item to be expelled is always marked as a non-sync write. This can
lead to a new sync write op for the same key being (erroneously) queued
in the same checkpoint. Any replicas being streamed items from this
checkpoint will then fail a precondition check after correctly
identifying that the new op would dedupe a sync write op.

As CheckpointCursors are always incremented before reading
an item they are not affected by the dummy value swap.

This patch invalidates the keyIndex entries before any splicing/value
swapping occurs.

Change-Id: I2de54fc62381d1af76e6f6e8fb15d4aa0dba596d
Reviewed-on: http://review.couchbase.org/116291
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>