Remove CheckpointIterator<>::getUnderlyingIterator()
This method was added to allow explicit access to the contained
iterator where needed to pass to stl methods.

Replacing it with an user-defined conversion operator allows
transparent usage where the underlying iterator is needed, without
adding an excessive amount of "surprise".

The access is (and was previously) const, so the internal state
shouldn't be accidentally corruptable.

Change-Id: I07e72c1b937b851a366261d6afa6006aba0350a8
Reviewed-on: http://review.couchbase.org/116294
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>