MB-36423: Don't reload un-persisted prepares during rollback
During rollback we reload the state of anything not yet persisted
from disk into the HashTable. This means that a prepare that has not
yet been persisted will be loaded into the HashTable during rollback.
This is incorrect behaviour as a completed prepare should not exist in
the HashTable for a persistent bucket and we know that the prepare on
disk has been completed as we would otherwise not have a different
unperisted one. This would then cause runloop exceptions when
replicating to this node.

Fix this by removing all unpersisted prepares from the HashTable
during rollback.

Change-Id: I2b7ea331aa649b6a38adb8087d043112530806eb
Reviewed-on: http://review.couchbase.org/116419
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>