Add setting to always collect trace info
Trace information was only collected iff the client asked for
it as it did affect latencies due to the extra calls to fetch
the system clock for adding the trace elements.

Unfortunately this means that one would need to change the
clients in order to get trace information in the "Slow operation"
entries in the logs. Not all customers needs to squeeze all
the juice out of the latencies, so it would be nice if one
could tell the server to collect the information and put
it in the logs whenever it encounters a slow command.

By setting "always_collect_trace_info" to true in the settings
(dynamic variable which may be toggled runtime) the server
will always collect trace information for operations (the
client won't get the timings back unless it asks for it)

Change-Id: Ic49e8914fd5cae557ff4f3bc7a2ba4108c41c0d2
Reviewed-on: http://review.couchbase.org/116485
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>