Slow down the logging during shutdown
In MB-36467 we've got a problem deleting a bucket, but due to
ns_server giving up waiting for memcached we end up in a
situation where memcached enters a tight loop trying to shut
down the connection and log every time. Given that we cycle
the log files and ns_server only keeps 20 of them (~400MB)
we'll only have a few seconds left.

Change-Id: I5320947a11321b9b5dcf886bf8fb1bb948e2bc13
Reviewed-on: http://review.couchbase.org/116503
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>