MB-36578: [BP] Dereference cookie when creating DCP conn with same name
Currently when we create a DCP connection with the same name as an
existing one we will disconnect the old one in favour of the new
one. This is fine, but we do not correctly deal with the old one.
We currently just call setDisconnect on but do not add it to
deadConnections. When we later call manageConnections we will not
find it in deadConnections and we will not decrement the refCount
of the cookie.

Ensure we do a proper disconnect of the connection by shutting down
tasks and adding the connection to deadConnections when we disconnect
a DCP connection due to it having the same name as a new one.

Change-Id: I2f84c74144d1c95f29b2505cacf1d02b87644e20
Reviewed-on: http://review.couchbase.org/116778
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>