MB-36227: Remove ConnHandler from vbConns outside of streamMutex
When seqno acking was added for SyncReplication we needed to add
the Consumer ConnHandler to vbConns so that we could ack the
correct Consumer efficiently. Due to a lock order inversion issue,
a change was made to promote the weak_ptr's stored in vbConns to
shared and then to ack the stream outside of the vbConnsLocks.
To remove the reliance on the element of vbConns being a
shared_ptr, fix this lock order inversion from the other side,
by removing elements from vbConns outside of the
Consumer::streamMutex and revert the original change.

The goal of this is to convert the ConnHandler element in vbConns
from a weak_ptr to a reference and avoid any unnecessary pointer
promotion as this was found to be very cache contentious for front
end ops as part of MB-33739. This change was made once before which
caused some crashes due to having invalid references lying around in
vbConns. As part of durability though we've found and fixed (hopefully)
all of the bugs where we kept invalid references in vbConns as they
were causing assertions to fire when seqno acking.

Change-Id: Ic76e75f59bf6d2bc8d0bfb53ad5081fd2696784c
Reviewed-on: http://review.couchbase.org/114679
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>